kafka/requirements.txt
psycopg2-binary==2.9.9
redis==5.0.8
httpx==0.27.2
pydantic==2.9.2
kafka-python==2.0.2 
google-genai==0.1.0